requirements.txt
Pillow==8.2.0
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1

numpy~=1.20.2
pygame~=2.0.1